Refactoring to make animations easier
Extracted the OnPageChangeCallback and PageTransformer into their own classes so I will be able to use fields for animations. Also good practice.
